requirements.txt
asgiref==3.3.4
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
psycopg2-binary==2.9.1
pytz==2021.1
requests==2.26.0
six==1.16.0
spotipy==2.19.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.2.0
